backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.0
aioredis>=2.0.0

# InfluxDB
influxdb-client>=1.40.0

# Modbus
pymodbus>=3.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Data processing
numpy>=1.26.0
scipy>=1.12.0
pydantic>=2.5.0
pyyaml>=6.0

# Gas properties (thermodynamics)
CoolProp>=6.4.1

# Utilities
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
celery>=5.3.0
flower>=2.0.0
